pyproject.toml
[tool.poetry]
name = "firmable-assignment"
version = "0.1.0"
description = "firmable assignment"
authors = ["Sagar Dobariya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
langchain = "^0.3.14"
openai = "^1.59.3"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
langchain-community = "^0.3.14"
langchain-openai = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
